+++ Simulator/ErrorHandling.h
@@ -3,7 +3,7 @@
 #include<string>
 using namespace std;
 
-int convertToInt(string s);
+long int convertToInt(string s);
 string seperateImmediate(string& s);
 bool IsvalidIandR(string s, int linenumber);
 bool IsValidImmediate(string s, bool flag);
